requirements.txt
easyocr==1.7.1
numpy==2.0.1
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==10.4.0
streamlit==1.37.1
ultralytics==8.2.77
